Struct libamadan::hillclimbing::Hillclimbing

pub struct Hillclimbing<S: Clone> { /* fields omitted */ }

Hillclimbing is the most basic metaheuristic. It starts with a random solution, mutates it slightly and if it achieves a better fitness after mutation than before, it replaces the initial solution. This is re-iterated until a given maximum number of iterations is reached or if a given maximum number of failed iterations is exceeded. This metaheuristic has no way of avoiding local maxima. If it gets stuck in a local maximum, it has no chance of ever leaving it again.

Methods

impl<S> Hillclimbing<S> where S: Clone

The implementation of the constructor for hillclimbing

fn new(max_iterations: u64, max_failed_iterations: u64) -> Hillclimbing<S>

Constructor for hillclimbing. Requires the maximum number of overall iterations and the maximum number of failed iterations as a parameter.