Trait libamadan::metaheuristic::Metaheuristic

pub trait Metaheuristic<S: Clone> {
    fn find(&self,
            rsg: &RandomSolutionGenerator<S>,
            fitness_function: &FitnessFunction<S>)
            -> S;
}

The Metaheuristic trait is used to generalize the interface of the different metaheuristics.

Required Methods

fn find(&self,
rsg: &RandomSolutionGenerator<S>,
fitness_function: &FitnessFunction<S>)
-> S

The actual metaheuristic optimization function. It finds a solution close to the optimimum, using a random solution generator and a function to evaluate a solution's fitness.